Function zng_wgt_inspector::debug::show_directional_query

pub fn show_directional_query(
    child: impl UiNode,
    orientation: impl IntoVar<Option<Orientation2D>>,
) -> impl UiNode

P Draw the directional query for closest sibling of the hovered focusable widget.

Window Only

This property only works if set in a window, if set in another widget it will log an error and not render anything.