zng_wgt_inspector
Module debug

Debug inspection properties.

Enums

InspectMode

Target of inspection properties.

Functions

show_bounds

P Draws a border for every target widget’s outer and inner bounds.

show_center_points

P Draws a debug dot in target widget’s center point.

show_directional_query

P Draw the directional query for closest sibling of the hovered focusable widget.

show_hit_test

P Draws the inner bounds that where tested for the mouse point.

show_rows

P Draws a border over every inlined widget row in the window.